# miscellaneous/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evently Booking Platform API",
...
    "contact": {
      "name": "Evently Support",
      "email": "droplakshya@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/events/popular/list": {
      "get": {
        "tags": ["events"],
        "summary": "Get Popular Events",
        "description": "Get popular events based on booking count.",
        "operationId": "get_popular_events_api_v1_events_popular_list_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Maximum number of results",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventResponse"
                  },
                  "title": "Response Get Popular Events Api V1 Events Popular List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/upcoming/list": {
      "get": {
        "tags": ["events"],
        "summary": "Get Upcoming Events",
        "description": "Get upcoming events with available capacity.",
        "operationId": "get_upcoming_events_api_v1_events_upcoming_list_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Maximum number of results",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventResponse"
                  },
                  "title": "Response Get Upcoming Events Api V1 Events Upcoming List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/waitlist/": {
      "post": {
        "tags": ["waitlist"],
        "summary": "Join Waitlist",
        "description": "Join the waitlist for a sold-out event.\n\n- **event_id**: ID of the event to join waitlist for\n- **requested_quantity**: Number of tickets requested",
        "operationId": "join_waitlist_api_v1_waitlist__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/waitlist/{event_id}": {
      "delete": {
        "tags": ["waitlist"],
        "summary": "Leave Waitlist",
        "description": "Leave the waitlist for an event.\n\n- **event_id**: ID of the event to leave waitlist for",
        "operationId": "leave_waitlist_api_v1_waitlist__event_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/waitlist/my-waitlist": {
      "get": {
        "tags": ["waitlist"],
        "summary": "Get My Waitlist Entries",
        "description": "Get current user's waitlist entries.\n\n- **status**: Optional filter by waitlist status\n- **limit**: Maximum number of entries to return (1-100)\n- **offset**: Number of entries to skip for pagination",
        "operationId": "get_my_waitlist_entries_api_v1_waitlist_my_waitlist_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaitlistStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by waitlist status",
              "title": "Status"
            },
            "description": "Filter by waitlist status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of entries to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum number of entries to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of entries to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of entries to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaitlistResponse"
                  },
                  "title": "Response Get My Waitlist Entries Api V1 Waitlist My Waitlist Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/waitlist/position/{event_id}": {
      "get": {
        "tags": ["waitlist"],
        "summary": "Get Waitlist Position",
        "description": "Get current user's position in the waitlist for an event.\n\n- **event_id**: ID of the event to check position for",
        "operationId": "get_waitlist_position_api_v1_waitlist_position__event_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/waitlist/admin/event/{event_id}": {
      "get": {
        "tags": ["waitlist"],
        "summary": "Get Event Waitlist",
        "description": "Get waitlist entries for a specific event (admin only).\n\n- **event_id**: ID of the event\n- **status**: Optional filter by waitlist status\n- **limit**: Maximum number of entries to return (1-500)\n- **offset**: Number of entries to skip for pagination",
        "operationId": "get_event_waitlist_api_v1_waitlist_admin_event__event_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaitlistStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by waitlist status",
              "title": "Status"
            },
            "description": "Filter by waitlist status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of entries to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of entries to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of entries to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of entries to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaitlistResponse"
                  },
                  "title": "Response Get Event Waitlist Api V1 Waitlist Admin Event  Event Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/waitlist/admin/stats/{event_id}": {
      "get": {
        "tags": ["waitlist"],
        "summary": "Get Waitlist Stats",
        "description": "Get waitlist statistics for an event (admin only).\n\n- **event_id**: ID of the event",
        "operationId": "get_waitlist_stats_api_v1_waitlist_admin_stats__event_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/waitlist/admin/notify/{event_id}": {
      "post": {
        "tags": ["waitlist"],
        "summary": "Notify Waitlist",
        "description": "Manually notify waitlisted users about available seats (admin only).\n\n- **event_id**: ID of the event\n- **available_quantity**: Number of seats that became available",
        "operationId": "notify_waitlist_api_v1_waitlist_admin_notify__event_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          },
          {
            "name": "available_quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Number of available seats",
              "title": "Available Quantity"
            },
            "description": "Number of available seats"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Event name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Event description"
          },
          "venue": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Venue",
            "description": "Event venue"
          },
          "event_date": {
            "type": "string",
            "format": "date-time",
            "title": "Event Date",
            "description": "Event date and time"
          },
          "total_capacity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Total Capacity",
            "description": "Total event capacity"
          },
          "price": {
            "type": "string",
            "title": "Price",
            "description": "Event ticket price"
          },
          "has_seat_selection": {
            "type": "boolean",
            "title": "Has Seat Selection",
            "description": "Whether event has seat selection",
            "default": false
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "available_capacity": {
            "type": "integer",
            "title": "Available Capacity"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "is_sold_out": {
            "type": "boolean",
            "title": "Is Sold Out"
          },
          "capacity_utilization": {
            "type": "number",
            "title": "Capacity Utilization"
          }
        },
        "type": "object",
        "required": [
          "name",
          "venue",
          "event_date",
          "total_capacity",
          "price",
          "id",
          "available_capacity",
          "is_active",
          "version",
          "created_at",
          "updated_at",
          "is_sold_out",
          "capacity_utilization"
        ],
        "title": "EventResponse",
        "description": "Schema for event response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "WaitlistCreate": {
        "properties": {
          "requested_quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Requested Quantity",
            "description": "Number of tickets requested"
          },
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id",
            "description": "ID of the event to join waitlist for"
          }
        },
        "type": "object",
        "required": ["requested_quantity", "event_id"],
        "title": "WaitlistCreate",
        "description": "Schema for creating a waitlist entry."
      },
      "WaitlistResponse": {
        "properties": {
          "requested_quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Requested Quantity",
            "description": "Number of tickets requested"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "position": {
            "type": "integer",
            "title": "Position",
            "description": "Position in the waitlist queue"
          },
          "status": {
            "$ref": "#/components/schemas/WaitlistStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "requested_quantity",
          "id",
          "user_id",
          "event_id",
          "position",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "WaitlistResponse",
        "description": "Schema for waitlist entry responses."
      },
      "WaitlistStatsResponse": {
        "properties": {
          "total_waitlisted": {
            "type": "integer",
            "title": "Total Waitlisted"
          },
          "active_waitlisted": {
            "type": "integer",
            "title": "Active Waitlisted"
          },
          "notified_count": {
            "type": "integer",
            "title": "Notified Count"
          },
          "converted_count": {
            "type": "integer",
            "title": "Converted Count"
          },
          "average_position": {
            "type": "number",
            "title": "Average Position"
          },
          "estimated_wait_time_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Time Hours",
            "description": "Estimated wait time in hours"
          }
        },
        "type": "object",
        "required": [
          "total_waitlisted",
          "active_waitlisted",
          "notified_count",
          "converted_count",
          "average_position"
        ],
        "title": "WaitlistStatsResponse",
        "description": "Schema for waitlist statistics."
      },
      "WaitlistStatus": {
        "type": "string",
        "enum": ["active", "notified", "expired", "converted"],
        "title": "WaitlistStatus",
        "description": "Enumeration for waitlist status."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}